chore: move GeminiRequestType file to entity folder
