--- src/main/java/com/gdg/kkia/gemini/GeminiRequestType.java
@@ -1,5 +1,0 @@
-package com.gdg.kkia.gemini;
-
-public enum GeminiRequestType {
-    chat, diary
-}
